improvement: adds fa spacer class
This very simple and neat hack adds CSS class which sets any icon to transparent color, thus with it, we have empty/blank fa icon, which can be used as an spacer and also it will maintain fixed width with other used icons.

As an example I've used it to replace spacer in customersearch form.